+++ scripts/test/Console/Command/Admin/Package/Upgrade.t
@@ -12,9 +12,63 @@ use utf8;
 
 use vars (qw($Self));
 
-my $CommandObject = $Kernel::OM->Get('Kernel::System::Console::Command::Admin::Package::Upgrade');
+my $Helper        = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
+my $PackageObject = $Kernel::OM->Get('Kernel::System::Package');
+my $ConfigObject  = $Kernel::OM->Get('Kernel::Config');
 
-my $ExitCode = $CommandObject->Execute();
+# Make sure to enable cloud services.
+$Helper->ConfigSettingChange(
+    Valid => 1,
+    Key   => 'CloudServices::Disabled',
+    Value => 0,
+);
+
+$Helper->ConfigSettingChange(
+    Valid => 1,
+    Key   => 'Package::AllowNotVerifiedPackages',
+    Value => 0,
+);
+
+my $RandomID = $Helper->GetRandomID();
+
+# Override Request() from WebUserAgent to always return some test data without making any
+#   actual web service calls. This should prevent instability in case cloud services are
+#   unavailable at the exact moment of this test run.
+my $CustomCode = <<"EOS";
+sub Kernel::Config::Files::ZZZZUnitTestAdminPackageManager${RandomID}::Load {} # no-op, avoid warning logs
+use Kernel::System::WebUserAgent;
+package Kernel::System::WebUserAgent;
+use strict;
+use warnings;
+## nofilter(TidyAll::Plugin::OTRS::Perl::TestSubs)
+{
+    no warnings 'redefine';
+    sub Request {
+        return (
+            Status  => '200 OK',
+            Content => '{"Success":1,"Results":{"PackageManagement":[{"Operation":"PackageVerify","Data":{"Test":"not_verified","TestPackageIncompatible":"not_verified"},"Success":"1"}]},"ErrorMessage":""},
+        );
+    }
+}
+1;
+EOS
+$Helper->CustomCodeActivate(
+    Code       => $CustomCode,
+    Identifier => 'Admin::Package::Upgrade' . $RandomID,
+);
+
+my $Location             = $ConfigObject->Get('Home') . '/scripts/test/sample/PackageManager/TestPackage.opm';
+my $UpgradeCommandObject = $Kernel::OM->Get('Kernel::System::Console::Command::Admin::Package::Upgrade');
+
+my $ExitCode = $UpgradeCommandObject->Execute($Location);
+
+$Self->Is(
+    $ExitCode,
+    1,
+    "Admin::Package::Upgrade exit code - package is not verified",
+);
+
+$ExitCode = $UpgradeCommandObject->Execute($Location);
 
 $Self->Is(
     $ExitCode,